i965/drm: Make brw_bo_alloc_tiled flags parameter 32-bit.
unsigned long is a terrible type for a bitfield - if you need fewer
than 32 bits, it wastes 4 bytes.  If you need more, things break on
32-bit builds.  Just use unsigned.

Even that's a bit ridiculous as we only have one flag today.
Still, it's at least somewhat better.

+++ src/mesa/drivers/dri/i965/brw_bufmgr.c
@@ -245,7 +245,7 @@ static struct brw_bo *
 bo_alloc_internal(struct brw_bufmgr *bufmgr,
                   const char *name,
                   uint64_t size,
-                  unsigned long flags,
+                  unsigned flags,
                   uint32_t tiling_mode,
                   uint32_t stride, uint64_t alignment)
 {
@@ -377,7 +377,7 @@ brw_bo_alloc(struct brw_bufmgr *bufmgr,
 struct brw_bo *
 brw_bo_alloc_tiled(struct brw_bufmgr *bufmgr, const char *name,
                    int x, int y, int cpp, uint32_t tiling,
-                   uint32_t *pitch, unsigned long flags)
+                   uint32_t *pitch, unsigned flags)
 {
    uint64_t size;
    uint32_t stride;
